--- week3/card-match/src/App.css
@@ -1,5 +1,0 @@
-body {
-  font-family: "Roboto", sans-serif;
-  text-align: center;
-  background-color: rgb(255, 249, 182);
-}
