+++ week3/card-match/src/main.jsx
@@ -2,15 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
 import { ThemeProvider, createGlobalStyle } from "styled-components";
-
-const darkTheme = {
-  textColor: "skyblue",
-  bgColor: "#111",
-};
+import { GlobalStyle } from "./styles/GlobalStyles.js";
+import theme from "./styles/theme.js";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <ThemeProvider theme={darkTheme}>
+    <GlobalStyle />
+    <ThemeProvider theme={theme}>
       <App />
     </ThemeProvider>
   </React.StrictMode>
