+++ week3/card-match/src/styles/theme.js
@@ -1,0 +1,15 @@
+const darkTheme = {
+    textColor: "skyblue",
+    bgColor: "#111",
+}
+
+const lightTheme = {
+    bgColor: "rgb(255, 249, 182)",
+}
+
+const theme = {
+    dark: darkTheme,
+    light: lightTheme,
+}
+
+export default theme
